# Conflicts:
#	dist/modules/autorisation.js
#	src/modules/autorisation.ts

+++ src/pages/chat/index.ts
@@ -1,25 +1,45 @@
 /// <reference path="../../../dist/modules/references.d.ts" />
 import Block from "../../../dist/modules/block.js";
 import getTemplateChat from "./chat.tmpl.js";
-import HTTPTransport from "../../../dist/modules/httpTransport.js";
+import { isValidMessage } from "../../../dist/modules/validation.js";
+import { getChats, getMessages } from "../../../dist/modules/autorization.js";
+import { getUserData, addChatToList, deleteChatFromList } from "../../modules/autorization.js";
+
+interface Message {
+    contact: string,
+    history: any[]
+}
+
+interface Contact {
+    contact: string,
+    pressed: string,
+    pianokey: string
+}
+
+interface Target {
+    messages: Message[],
+    chat: Chat
+};
+
+type strElement = string | null | undefined;
 
 export class Chat extends Block {
-    messages: [{ contact: string, history: any[] }];
-    contacts: [{ contact: string, pressed: string, pianokey: string }];
+    messages: Message[];
+    contacts: Contact[];
     fileImage: string;
     fileImageSound: string;
     name: string;
     errorMessage: string;
     proxyData: any;
 
-    constructor(props) {
+    constructor(props: Object) {
         super("chat", props);
         this.fileImage = "";
         this.name = "";
         this.errorMessage = "";
         this.getFileData();
 
-        const data = {
+        const data: Target = {
             messages: this.messages,
             chat: this
         };
@@ -28,87 +48,62 @@ export class Chat extends Block {
                 const value = target[prop];
                 return typeof value === "function" ? value.bind(target) : value;
             },
-            set(target, prop, value) {
+            async set(target, prop, value) {
                 target[prop] = value;
-                target.chat.render();
-                target.chat.setEvents();
+                await target.chat.render();
+                //target.chat.setEvents();
                 return true;
             },
         });
     }
     private async getFileData() {
-        await this.getMessages();
-        await this.getPianokeys();
+        this.messages = await getMessages();
+        this.contacts = await getChats();
+        const userData = await getUserData();
+        this.name = userData.name;
+        this.fileImage = userData.avatar;
     }
 
     getAllData() {
-        let messages = [];
-        const pianoKeys = this.contacts;
-        const fileImage = this.getImage();
-        const fileAdd  = "../../../data/img/add.jpg";
+        const fileAdd = "../../../data/img/add.jpg";
         const fileDelete = "../../../data/img/delete.jpg";
         const fileImageSound = "../../../data/img/sound.jpg";
-        let name = this.getName();
-        let errorMessage = this.getErrorMessage();
-        return { pianoKeys: pianoKeys, messages: messages, fileImage: fileImage, fileImageSound: fileImageSound, name: name, 
-            errorMessage: errorMessage, fileAdd: fileAdd, fileDelete: fileDelete };
+        let name = this.name
+        let errorMessage = "Пустое сообщение";
+        return {
+            pianoKeys: this.contacts, messages: this.messages, fileImage: this.fileImage, fileImageSound: fileImageSound, name: name,
+            errorMessage: errorMessage, fileAdd: fileAdd, fileDelete: fileDelete
+        };
     }
 
     getChatMessages() {
         return this.messages;
     }
 
-    private async getMessages() {
-        const httpTransport = new HTTPTransport;
-        const res = await httpTransport.get('../../../data/messages.json');
-        const resHTTP = await JSON.parse(res.response);
-        this.proxyData.messages = resHTTP;
-        this.messages = resHTTP;
-    }
-
-    private async getPianokeys() {
-        const httpTransport = new HTTPTransport;
-        const res = await httpTransport.get('../../../data/contacts.json');
-        const resHTTP = await JSON.parse(res.response);
-        this.contacts = resHTTP;
-        this.proxyData.contacts = resHTTP;
-    }
-
-    private getImage() {
-        const result = "../../../data/img/ava.png";
-        return result;
-    }
-
-    private getName() {
-        const result = "Анастасия";
-        return result;
-    }
-
-    private getErrorMessage() {
-        const result = "Пустое сообщение";
-        return result;
-    }
-
-    render(context = undefined) {
-        if (context === undefined)
+    async render(context: Object | undefined = undefined) {
+        if (context === undefined) {
+            await this.getFileData();
             context = this.getAllData();
+        }
+
         const result = compileTemplate('.app', getTemplateChat(), context);
+
         return result;
     }
 
-    showChatHistory(elementContact) {
-        const idContact = this.getIdContact(elementContact);
+    async showChatHistory(elementContact: HTMLElement) {
+        const idContact: strElement = this.getIdContact(elementContact);
         if (idContact !== undefined) {
             const history = this.getHistoryChat(idContact);
             this.updateContacts(idContact);
             const context = this.getAllData();
             context.messages = history;
-            this.render(context);
+            await this.render(context);
             this.setEvents();
         }
     }
 
-    private updateContacts(idContact) {
+    private updateContacts(idContact: strElement) {
         for (let i = 0; i < this.contacts.length; i++) {
             if (this.contacts[i].contact.trim() == idContact.trim()) {
                 this.contacts[i].pressed = "chat-wrapper__white-pianokey-pressed";
@@ -122,7 +117,7 @@ export class Chat extends Block {
         }
     }
 
-    getIdContact(elementContact) {
+    getIdContact(elementContact: HTMLElement): strElement {
         let element = undefined;
         let result = undefined;
 
@@ -140,7 +135,7 @@ export class Chat extends Block {
         return result;
     }
 
-    getHistoryChat(idContact) {
+    getHistoryChat(idContact: strElement): Message[] | undefined {
         for (let i = 0; i < this.messages.length; i++) {
             let elem = this.messages[i];
             if ((elem.contact != null) && (elem.contact != undefined) && (elem.contact.trim() == idContact.trim()))
@@ -152,21 +147,23 @@ export class Chat extends Block {
 
     setEvents() {
         const nameHiddenError = "chat-wrapper";
-        const inputs = [{ input: "ineditor", value: validateMessage("ineditor", nameHiddenError) }];
-        const button = document.getElementById("ineditor");
-        button.addEventListener('keypress', function (e) {
-            if (e.key === 'Enter') {
-                if (isValidValues(inputs, nameHiddenError)) {
-                    let data = getData(inputs);
-                    console.log(data);
+        const inputs = [{ input: "ineditor", value: isValidMessage }];
+        const button = <HTMLElement>document.getElementById("ineditor");
+        
+        if (button !== null) {
+            button.addEventListener('keypress', function (e) {
+                if (e.key === 'Enter') {
+                    if (isValidValues(inputs, nameHiddenError)) {
+                        let data = getData(inputs);
+                        console.log(data);
+                    }
+                } else {
+                    let elementError = document.getElementById('err-ineditor7');
+                    if (elementError && (!elementError.classList.contains(nameHiddenError)))
+                        elementError.classList.add(nameHiddenError);
                 }
-            } else {
-                let elementError = document.getElementById('err-ineditor7');
-                if (elementError && (!elementError.classList.contains(nameHiddenError)))
-                    elementError.classList.add(nameHiddenError);
-            }
-        });
-
+            });
+        };
         const elementButtonId = document.querySelector(".chat-wrapper__left-part");
 
         if (elementButtonId != null) {
@@ -178,45 +175,50 @@ export class Chat extends Block {
             );
         }
         const elementSound = <HTMLImageElement>document.getElementById("sound");
-        elementSound.addEventListener('click', function (e) {
-            if (elementSound.src == "http://mf.messenger.praktikum.yandex/data/img/sound.jpg") {
-                elementSound.src = "http://mf.messenger.praktikum.yandex/data/img/nosound.jpg";
-                (<any>window).noSound = true;
-            }
-            else {
-                elementSound.src = "http://mf.messenger.praktikum.yandex/data/img/sound.jpg";
-                (<any>window).noSound = false;
-            }
-        });
+        if (elementSound != null) {
+            elementSound.addEventListener('click', function (e) {
+                if (elementSound.src == "http://mf.messenger.praktikum.yandex/data/img/sound.jpg") {
+                    elementSound.src = "http://mf.messenger.praktikum.yandex/data/img/nosound.jpg";
+                    (<any>window).noSound = true;
+                }
+                else {
+                    elementSound.src = "http://mf.messenger.praktikum.yandex/data/img/sound.jpg";
+                    (<any>window).noSound = false;
+                }
+            })
+        };
+
         const elementAddChat = <HTMLElement>document.querySelector("#add");
-        elementAddChat.addEventListener('click', function(e) { addChat() });
+        if (elementAddChat != null)
+            elementAddChat.addEventListener('click', function (e) { addChat() });
         const elementDeleteChat = <HTMLElement>document.querySelector("#delete");
-        elementDeleteChat.addEventListener('click', function(e) { deleteChat() });
+        if (elementDeleteChat != null)
+            elementDeleteChat.addEventListener('click', function (e) { deleteChat() });
     }
 }
 
-function addChat() {
+async function addChat() {
     const elementSearch = <HTMLInputElement>document.querySelector("#search");
     const value = elementSearch.value;
     if (value === "")
         alert("Укажите имя чата")
     else {
         const curChat = (<any>window).currentBlock;
-        curChat.contacts.unshift({ "contact": value, "countmes": "", "hidden": false, "pressed": false, "pianokey": "7si.mp3" });
-        curChat.render();
+        addChatToList({ "contact": value, "countmes": "", "hidden": false, "pressed": false, "pianokey": "7si.mp3" });
+        await curChat.render();
         curChat.setEvents();
     }
 }
 
-function deleteChat() {
+async function deleteChat() {
     const curChat = (<any>window).currentBlock;
     const idContact = getCurrentContact(curChat.contacts);
-    curChat.contacts.splice(idContact, 1);
-    curChat.render();
+    deleteChatFromList(curChat.contacts[idContact]);
+    await curChat.render();
     curChat.setEvents();
 }
 
-function getCurrentContact(contacts) {
+function getCurrentContact(contacts: Contact[]) {
 
     for (let i = 0; i < contacts.length; i++) {
         if (contacts[i].pressed)
@@ -226,9 +228,3 @@ function getCurrentContact(contacts) {
     return -1;
 }
 
-
-function render(query, block) {
-    const root = document.querySelector(query);
-    root.appendChild(block.getContent());
-    return root;
-}
